refactor(chat): extract role/application event renderers from MessageDisplay
Stage 4b (1/2). Moves the 15 role- and application-event render* functions
into a new messageEventRenderers.jsx via a createEventRenderers(ctx) factory:
ctx carries the parent-owned mention components and helpers, destructured once,
with the function bodies kept verbatim. MessageDisplay 4194 → 3365 lines.
Membership/invitation/ownership renderers follow in the next commit.
